Update includedCovariateIds to use Long instead of Integer (#18)

+++ src/main/java/org/ohdsi/analysis/featureextraction/design/CovariateSettings.java
@@ -806,7 +806,7 @@ public interface CovariateSettings extends RLangClass {
      * @return
      */
     @JsonGetter("includedCovariateIds")
-    List<Integer> getIncludedCovariateIds();
+    List<Long> getIncludedCovariateIds();
 
     /**
      *
